Move theme response types out of swap utilities
