+++ src/features/charts/polymarket/stores/polymarketChartStore.ts
@@ -1,12 +1,12 @@
 import { createQueryStore } from '@storesjs/stores';
 
-import { type ResponseByTheme } from '@/__swaps__/utils/swaps';
 import { POLYMARKET_SPORTS_MARKET_TYPE } from '@/features/polymarket/constants';
 import { usePolymarketEventStore } from '@/features/polymarket/stores/polymarketEventStore';
 import { getOutcomeTeamColor } from '@/features/polymarket/utils/getOutcomeTeam';
 import { isThreeWayMoneyline } from '@/features/polymarket/utils/marketClassification';
 import { isDrawMarket } from '@/features/polymarket/utils/sports';
 import { time } from '@/framework/core/utils/time';
+import { type ResponseByTheme } from '@/theme/types';
 
 import { fetchPriceHistory } from '../api/clobClient';
 import { fetchGammaEvent } from '../api/gammaClient';
