+++ src/features/polymarket/utils/getMarketColor.ts
@@ -1,4 +1,4 @@
-import { getColorValueForThemeWorklet, type ResponseByTheme } from '@/__swaps__/utils/swaps';
+import { getColorValueForThemeWorklet } from '@/__swaps__/utils/swaps';
 import { palettes } from '@/design-system/color/palettes';
 import { type PolymarketPosition, type PolymarketTeamInfo } from '@/features/polymarket/types';
 import { type PolymarketMarket, type RawPolymarketMarket } from '@/features/polymarket/types/polymarket-event';
@@ -7,6 +7,7 @@ import { getOutcomeTeamColor } from '@/features/polymarket/utils/getOutcomeTeam'
 import { isTeamBasedOutcome } from '@/features/polymarket/utils/marketClassification';
 import { isDrawMarket } from '@/features/polymarket/utils/sports';
 import { getHighContrastColor } from '@/hooks/useAccountAccentColor';
+import { type ResponseByTheme } from '@/theme/types';
 
 type MarketColors = {
   color: ResponseByTheme<string>;
