+++ src/theme/types.ts
@@ -1,0 +1,4 @@
+export type ResponseByTheme<T> = {
+  dark: T;
+  light: T;
+};
